Add task finding feature
