+++ src/main/java/duke/commands/FindCommand.java
@@ -1,0 +1,16 @@
+package duke.commands;
+
+/**
+ * Lists all tasks in the task list.
+ */
+public class FindCommand extends Command {
+    private String keyword;
+
+    public FindCommand(String keyword) {
+        this.keyword = keyword;
+    }
+
+    public String execute() {
+        return super.taskList.findTasks(this.keyword);
+    }
+}
